requirements-web.txt
opencv-python-headless>=4.8.1.78
numpy>=1.24.3
matplotlib>=3.7.2
flask>=3.0.0
flask-cors>=4.0.0
pyyaml>=6.0.1
pytest>=7.4.3
pytest-cov>=4.1.0
gunicorn>=21.2.0
pillow>=10.1.0
